+++ src/components/sync/SyncOrchestration.tsx
@@ -1,19 +1,15 @@
-import { AnnouncementSync } from './AnnouncementSync';
 import { BandSync } from './BandSync';
 import { CacheVersionCheck } from './CacheVersionCheck';
 import { DuckNotificationsListener } from './DuckNotificationsListener';
-import { DuckSync } from './DuckSync';
-import { PickSync } from './PickSync';
 import { PushSetup } from './PushSetup';
+import { ReconnectSync } from './ReconnectSync';
 
 export function SyncOrchestration() {
   return (
     <>
       <CacheVersionCheck />
       <BandSync />
-      <PickSync />
-      <AnnouncementSync />
-      <DuckSync />
+      <ReconnectSync />
       <PushSetup />
       <DuckNotificationsListener />
     </>